refactor ProtocolFilterWrapper$1 to FilterNode
